pyproject.toml
[project]
name = "nvr-search"
version = "1.0.0"
description = "NVR 영상 VLM 캡셔닝 및 자연어 검색 시스템"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.0",
    "chromadb>=0.6",
    "sentence-transformers>=3.0",
    "opencv-python>=4.9",
    "watchdog>=4.0",
    "pyyaml>=6.0",
    "pyinstaller>=6.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
]
